implementing tokens and Scedule class to ensure that a user will have 2 token for each day and 1 delete token for each month

+++ src/main/java/dev/yassiraitelghari/web/DeleteTaskServlet.java
@@ -1,5 +1,6 @@
 package dev.yassiraitelghari.web;
 
+import dev.yassiraitelghari.domain.Task;
 import dev.yassiraitelghari.repositories.TaskRepository;
 import dev.yassiraitelghari.repositories.TaskRepositoryImp;
 import dev.yassiraitelghari.services.TagService;
@@ -15,13 +16,25 @@ import java.io.IOException;
 public class DeleteTaskServlet extends HttpServlet {
     private TaskService taskService = new TaskServiceImp();
     private TagService tagService = new TagServiceImp();
+
     @Override
-    protected  void doPost(HttpServletRequest request , HttpServletResponse response) throws IOException {
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
         String id = request.getParameter("task_id");
         int parsedId = Integer.parseInt(id);
-        if(tagService.deleteByTask(parsedId)){
-            taskService.delete(parsedId);
+        Task task = taskService.findTask(parsedId);
+        if (task.isAddedByMe()) {
+            if (tagService.deleteByTask(parsedId)) {
+                taskService.delete(parsedId);
+            }
+        } else {
+            if(task.getUser().getDeleteJeton()>0){
+                if (tagService.deleteByTask(parsedId)) {
+                    taskService.delete(parsedId);
+                }
+                //must make delete count 0
+            }
         }
+
         response.sendRedirect("../dashboard");
     }
 }
